+++ apps/web/e2e/data/users.ts
@@ -1,8 +1,7 @@
 export const testUsers = {
 	registered: {
 		mail: 'user1@test.com',
-		password: 'qwerty123',
-		id: 'b67cd35d-1e25-418a-a68b-afa0aaaabe36'
+		password: 'qwerty123'
 	},
 	new: {
 		mail: 'correct@mail.com',
