+++ apps/web/e2e/pages/auth/signin/signin-form.page.ts
@@ -2,19 +2,19 @@ import type { Locator, Page } from '@playwright/test';
 
 export class SignInPage {
 	private readonly page: Page;
-	private readonly pageLocator: Locator;
+	private readonly _page: Locator;
 	private readonly emailInput: Locator;
 	private readonly passwordInput: Locator;
-	private readonly submitButton: Locator;
 	private readonly signUpLink: Locator;
+	public readonly submitButton: Locator;
 
 	constructor(page: Page) {
 		this.page = page;
-		this.pageLocator = page.getByTestId('sign-in-form');
-		this.emailInput = this.pageLocator.getByTestId('email');
-		this.passwordInput = this.pageLocator.getByTestId('password');
-		this.submitButton = this.pageLocator.getByTestId('submit');
-		this.signUpLink = this.pageLocator.getByTestId('link-to-signup');
+		this._page = this.page.getByTestId('sign-in-form');
+		this.emailInput = this._page.getByTestId('email');
+		this.passwordInput = this._page.getByTestId('password');
+		this.submitButton = this._page.getByTestId('submit');
+		this.signUpLink = this._page.getByTestId('link-to-signup');
 	}
 
 	async fillForm(email: string, password: string) {
