+++ apps/web/src/routes/(app)/menus/+page.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+	import Button from '$lib/components/Button.svelte';
 	import Menu from '$lib/components/Menu.svelte';
 
 	export let data;
@@ -6,23 +7,18 @@
 
 <div class="flex flex-col items-center mx-5" data-testid="menu-page">
 	<h1 class="text-3xl font-bold my-3 mb-8">Your Menus</h1>
+
+	{#if data.menus.length === 0}
+		<p class="mb-2">You don't have any menu yet.</p>
+	{/if}
+
 	{#if data.menus.length > 0}
 		{#each data.menus as menu}
 			<Menu {menu} />
 		{/each}
-		<form method="POST" action="?/createMenu">
-			<button
-				class="p-1 my-3 px-3 border-2 text-center border-slate-200 shadow-md rounded"
-				data-testid="new-menu">Create a menu</button
-			>
-		</form>
-	{:else}
-		<p class="mb-2">You don't have any menu yet.</p>
-		<form method="POST" action="?/createMenu">
-			<button
-				class="p-1 my-3 px-3 border-2 text-center border-slate-200 shadow-md rounded"
-				data-testid="new-menu">Create a new one</button
-			>
-		</form>
 	{/if}
+
+	<form method="POST" action="?/createMenu">
+		<Button label="New menu" id="new-menu" />
+	</form>
 </div>
